estimateTransferRatioForScript just for LegacyJavascript
